+++ src/components/highscores/highscores-component.css
@@ -7,6 +7,7 @@
   font-size: large;
   text-align: left;
 }
+
 .highscores th {
   font-weight: bold;
   font-size: large;
